+++ src/main/kotlin/racingcar/view/output/PrintOutput.kt
@@ -1,4 +1,4 @@
-package racingcar.view.result
+package racingcar.view.output
 
 class PrintOutput : Output {
     override fun show(message: String) {
