+++ src/test/kotlin/racingcar/StubOutput.kt
@@ -1,0 +1,11 @@
+package racingcar
+
+import racingcar.view.output.Output
+
+class StubOutput : Output {
+    lateinit var result: String
+
+    override fun show(message: String) {
+        this.result = message
+    }
+}
